requirements.txt
selenium==3.141.0
pytest==5.4.3
requests==2.24.0
pytest-xdist==1.32.0
allure-pytest==2.8.16
mariadb~=1.0.0